serializers: fixing / updates to dublincore * added location to export + tests * closes https://github.com/inveniosoftware/invenio-rdm-records/issues/1201
